+++ django/librarian/utils/process_engine.py
@@ -18,7 +18,8 @@ from bs4 import BeautifulSoup
 from markdownify import markdownify
 from structlog import get_logger
 
-from librarian.utils.extract_zip import extract_file
+from librarian.models import Document, SavedFile
+from librarian.utils.extract_zip import process_zip_file
 from librarian.utils.markdown_splitter import MarkdownSplitter
 from otto.models import Cost
 
@@ -121,7 +122,7 @@ def guess_content_type(
         "application/pdf",
         "application/xml",
         "application/vnd.ms-outlook",
-        "application/zip",
+        "application/x-zip-compressed",
         "text/html",
         "text/markdown",
         "text/csv",
@@ -154,7 +155,7 @@ def guess_content_type(
             return "application/vnd.ms-outlook"
 
         if path.endswith(".zip"):
-            return "application/zip"
+            return "application/x-zip-compressed"
         # Use filetype library to guess the content type
         kind = filetype.guess(content)
         if kind and not path.endswith(".md"):
@@ -191,7 +192,7 @@ def get_process_engine_from_type(type):
         return "POWERPOINT"
     elif "application/vnd.ms-outlook" in type:
         return "OUTLOOK_MSG"
-    elif "application/zip" in type:
+    elif "application/x-zip-compressed" in type:
         return "ZIP"
     elif "application/pdf" in type:
         return "PDF"
@@ -269,7 +270,7 @@ def extract_markdown(
             md = msg_to_markdown(content)
         elif process_engine == "ZIP":
             enable_markdown = False
-            md = extract_file(content, data_source_id)
+            md = process_zip_file(content, data_source_id)
         elif process_engine == "CSV":
             md = csv_to_markdown(content)
         elif process_engine == "EXCEL":
@@ -752,3 +753,35 @@ def resize_to_azure_requirements(content):
             image.save(output, format="PNG")
             content = output.getvalue()
             return content
+
+
+def process_file(file, data_source_id, name, content_type):
+    from librarian.utils.process_engine import generate_hash
+
+    file_hash = generate_hash(file.read())
+    file_exists = SavedFile.objects.filter(sha256_hash=file_hash).exists()
+    if file_exists:
+        file_obj = SavedFile.objects.filter(sha256_hash=file_hash).first()
+        logger.info(
+            f"Found existing SavedFile for {file.name}", saved_file_id=file_obj.id
+        )
+        # Check if identical document already exists in the DataSource
+        existing_document = Document.objects.filter(
+            data_source_id=data_source_id,
+            filename=file.name,
+            file__sha256_hash=file_hash,
+        ).first()
+        # Skip if filename and hash are the same, and processing status is SUCCESS
+        if existing_document:
+            if existing_document.status != "SUCCESS":
+                existing_document.process()
+            return
+    else:
+        file_obj = SavedFile.objects.create(content_type=content_type)
+        file_obj.file.save(name, file)
+        file_obj.generate_hash()
+
+    document = Document.objects.create(
+        data_source_id=data_source_id, file=file_obj, filename=name
+    )
+    document.process()
